buildGqlQuery typings
