+++ package.json
@@ -41,6 +41,7 @@
   },
   "devDependencies": {
     "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.179",
     "graphql": "^14.5.8",
     "husky": "^4.3.0",
     "jest": "^27.5.1",
