+++ src/typings.d.ts
@@ -1,0 +1,2 @@
+declare module 'graphql-ast-types-browser';
+declare module 'ra-data-graphql';
